Split composer slash command helpers
